LayoutTests/js/typedarray-reduce-expected.txt

This test checks the behavior of the TypedArray.prototype.reduce function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.reduce.length is 1
PASS Int32Array.prototype.reduce.name is 'reduce'
PASS isSameFunctionForEachTypedArrayPrototype('reduce') is true
PASS testPrototypeReceivesArray('reduce', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('reduce', '(createArray)', [12, 5, 8, 13, 44], [12, 5, 8, 13, 44], [12, 5, 8, 13, 44]) is true
PASS testPrototypeFunction('reduce', '(sum)', [1, 2, 3, 4, 5], 15) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('reduce', '(createArray, [1])', [12, 23, 11, 1, 45], [1, 12, 23, 11, 1, 45], [12, 23, 11, 1, 45]) is true

3.0 Array Element Changing
PASS testPrototypeFunction('reduce', '(createArrayAndChange, [])', [12, 15, 2, 13, 44], [12, 15, 2, 5, 5], [5, 5, 5, 5, 5]) is true

4.0 Exception Test
PASS testPrototypeFunction('reduce', '(isBigEnoughAndException)', [12, 15, 10, 13, 44], false) threw exception exception from function.

5.0 Wrong Type for Callback Test
PASS testPrototypeFunction('reduce', '(8)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduce callback must be a function.
PASS testPrototypeFunction('reduce', '("wrong")', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduce callback must be a function.
PASS testPrototypeFunction('reduce', '(new Object())', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduce callback must be a function.
PASS testPrototypeFunction('reduce', '(null)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduce callback must be a function.
PASS testPrototypeFunction('reduce', '()', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduce callback must be a function.

PASS successfullyParsed is true

TEST COMPLETE
